+++ plugin/brdrq/test/test_autocorrectborders.py
@@ -30,7 +30,6 @@ class TestAutoCorrectBorders(unittest.TestCase):
     def tearDown(self):
         QGISAPP.processingRegistry().removeProvider(self.provider)
         QgsProject.instance().clear()
-
     def test_autocorrectborders(self):
         # See https://gis.stackexchange.com/a/276979/4972 for a list of algorithms
         foldername = QgsProcessingParameterFolderDestination(name="brdrQ").generateTemporaryDestination()
@@ -52,7 +51,9 @@ class TestAutoCorrectBorders(unittest.TestCase):
                 "WORK_FOLDER": foldername,
                 "ENUM_OD_STRATEGY": 1,
                 "THRESHOLD_OVERLAP_PERCENTAGE": 50,
+                "REVIEW_PERCENTAGE": 10,
                 "ADD_FORMULA": True,
+                "STABILITY": True,
                 "ADD_ATTRIBUTES": True,
                 "SHOW_INTERMEDIATE_LAYERS": True,
                 "PREDICTIONS": False,
@@ -60,7 +61,43 @@ class TestAutoCorrectBorders(unittest.TestCase):
             },
         )
         featurecount = layer_theme.featureCount()
-        assert len(output)==4
+        assert len(output)==5
+        for o in output.values():
+            assert isinstance(o,QgsVectorLayer)
+            assert o.featureCount()==featurecount
+
+    def test_autocorrectborders_fiscal(self):
+        # See https://gis.stackexchange.com/a/276979/4972 for a list of algorithms
+        foldername = QgsProcessingParameterFolderDestination(name="brdrQ").generateTemporaryDestination()
+
+        path = os.path.join(os.path.dirname(__file__), "themelayer_test.geojson")
+        themelayername = "themelayer_test"
+        layer_theme = QgsVectorLayer(path, themelayername)
+        QgsProject.instance().addMapLayer(layer_theme)
+
+        output = processing.run(
+            "brdrqprovider:brdrqautocorrectborders",
+            {
+                "INPUT_THEMATIC": themelayername,
+                "COMBOBOX_ID_THEME": "theme_identifier",
+                "RELEVANT_DISTANCE": 2,
+                "ENUM_REFERENCE": 30,  # fiscal2023
+                "INPUT_REFERENCE": None,
+                "COMBOBOX_ID_REFERENCE": "",
+                "WORK_FOLDER": foldername,
+                "ENUM_OD_STRATEGY": 1,
+                "THRESHOLD_OVERLAP_PERCENTAGE": 50,
+                "REVIEW_PERCENTAGE": 10,
+                "ADD_FORMULA": True,
+                "STABILITY": True,
+                "ADD_ATTRIBUTES": True,
+                "SHOW_INTERMEDIATE_LAYERS": True,
+                "PREDICTIONS": False,
+                "SHOW_LOG_INFO": False,
+            },
+        )
+        featurecount = layer_theme.featureCount()
+        assert len(output)==5
         for o in output.values():
             assert isinstance(o,QgsVectorLayer)
             assert o.featureCount()==featurecount
@@ -99,7 +136,9 @@ class TestAutoCorrectBorders(unittest.TestCase):
                 "WORK_FOLDER": foldername,
                 "ENUM_OD_STRATEGY": 1,
                 "THRESHOLD_OVERLAP_PERCENTAGE": 50,
+                "REVIEW_PERCENTAGE": 10,
                 "ADD_FORMULA": True,
+                "STABILITY": True,
                 "ADD_ATTRIBUTES": True,
                 "SHOW_INTERMEDIATE_LAYERS": True,
                 "PREDICTIONS": False,
@@ -107,7 +146,7 @@ class TestAutoCorrectBorders(unittest.TestCase):
             },
         )
         featurecount = layer_theme.featureCount()
-        assert len(output)==4
+        assert len(output)==5
         for o in output.values():
             assert isinstance(o,QgsVectorLayer)
             assert o.featureCount()==featurecount
@@ -138,7 +177,9 @@ class TestAutoCorrectBorders(unittest.TestCase):
                 "WORK_FOLDER": foldername,
                 "ENUM_OD_STRATEGY": 1,
                 "THRESHOLD_OVERLAP_PERCENTAGE": 50,
+                "REVIEW_PERCENTAGE": 10,
                 "ADD_FORMULA": False,
+                "STABILITY": True,
                 "ADD_ATTRIBUTES": True,
                 "SHOW_INTERMEDIATE_LAYERS": True,
                 "PREDICTIONS": False,
@@ -146,7 +187,7 @@ class TestAutoCorrectBorders(unittest.TestCase):
             },
         )
         featurecount = layer_theme.featureCount()
-        assert len(output)==4
+        assert len(output)==5
         for o in output.values():
             assert isinstance(o,QgsVectorLayer)
             assert o.featureCount()==featurecount
@@ -170,7 +211,9 @@ class TestAutoCorrectBorders(unittest.TestCase):
                 "WORK_FOLDER": foldername,
                 "ENUM_OD_STRATEGY": 1,
                 "THRESHOLD_OVERLAP_PERCENTAGE": 50,
+                "REVIEW_PERCENTAGE": 10,
                 "ADD_FORMULA": True,
+                "STABILITY": True,
                 "ADD_ATTRIBUTES": True,
                 "SHOW_INTERMEDIATE_LAYERS": True,
                 "PREDICTIONS": True,
@@ -178,7 +221,7 @@ class TestAutoCorrectBorders(unittest.TestCase):
             },
         )
         featurecount = layer_theme.featureCount()
-        assert len(output)==4
+        assert len(output)==5
         for o in output.values():
             assert isinstance(o,QgsVectorLayer)
             assert o.featureCount()>=featurecount
@@ -204,7 +247,9 @@ class TestAutoCorrectBorders(unittest.TestCase):
                     "WORK_FOLDER": foldername,
                     "ENUM_OD_STRATEGY": 8,
                     "THRESHOLD_OVERLAP_PERCENTAGE": 50,
+                    "REVIEW_PERCENTAGE": 10,
                     "ADD_FORMULA": True,
+                    "STABILITY": True,
                     "ADD_ATTRIBUTES": True,
                     "SHOW_INTERMEDIATE_LAYERS": True,
                     "PREDICTIONS": False,